types: review fixes and suggestions
